cleaned up code, getting mysterious errors from slashing
